requirements.txt
pytest==8.3.4
playwright==1.49.1
allure-pytest==2.13.5
python-dotenv==1.0.1
tzdata==2025.2
